% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notion-package.R
\docType{package}
\name{notion-package}
\alias{notion}
\alias{notion-package}
\title{notion: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EthanSansom/notion}
  \item Report bugs at \url{https://github.com/EthanSansom/notion/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
